fix: test and troubleshoot
